mcp_github/requirements.txt
# MCP Core Dependencies
mcp>=1.0.0
pydantic>=2.0.0

# HTTP Client
aiohttp>=3.8.0
aiofiles>=23.0.0

# GitHub API
PyGithub>=1.59.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
asyncio-mqtt>=0.16.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0

# Optional: For enhanced features
cryptography>=41.0.0
PyJWT>=2.8.0
marshmallow>=3.20.0 